Add MainFrame component to handle session verification and ChatList and Header rendering
